Fizes for supporting disabled private network

+++ src/hetzner/network.cr
@@ -3,6 +3,6 @@ require "json"
 class Hetzner::Network
   include JSON::Serializable
 
-  property id : Int32
+  property id : Int64
   property name : String
 end
